+++ packages/sif-common-forms-ds/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @navikt/sif-common-forms-ds
 
+## 4.1.2
+
+### Patch Changes
+
+-   41aaa80: Ta i bruk countryIsMemberOfEøsOrEfta fra sif-common-formik
+-   Updated dependencies [41aaa80]
+    -   @navikt/sif-common-core-ds@6.1.2
+
+## 4.1.1
+
+### Patch Changes
+
+-   @navikt/sif-common-core-ds@6.1.1
+
 ## 4.1.0
 
 ### Minor Changes
@@ -9,7 +23,7 @@
 ### Patch Changes
 
 -   Updated dependencies [908bb770]
-    -   @navikt/sif-common-core-ds@7.0.0
+    -   @navikt/sif-common-core-ds@6.1.0
 
 ## 4.0.3