+++ packages/sif-common-soknad-ds/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @navikt/sif-common-soknad-ds
 
+## 6.1.1
+
+### Patch Changes
+
+-   Updated dependencies [94686aa]
+    -   @navikt/sif-common-formik-ds@0.13.0
+
 ## 6.1.0
 
 ### Minor Changes